Implement EVM deployment & proof submission using ethers
