+++ examples/evm-verifier.rs
@@ -1,273 +1,24 @@
-use halo2_proofs::{
-    circuit::{Layouter, SimpleFloorPlanner, Value},
-    dev::MockProver,
-    halo2curves::{
-        bn256::{Bn256, Fq, Fr, G1Affine},
-        ff::Field,
+use eyre::Result;
+use halo2_proofs::{halo2curves::bn256::Fr, plonk::Circuit};
+use rand::rngs::OsRng;
+use std::{env, path::Path};
+use zero_g::{
+    checked_in_test_data::*,
+    eth::{
+        deploy::{evm_deploy, evm_verify},
+        gen_evm_verifier, gen_pk, gen_proof, gen_srs,
+        vanilla_plonk_circuit::StandardPlonk,
     },
-    plonk::{
-        create_proof, keygen_pk, keygen_vk, verify_proof, Advice, Circuit, Column,
-        ConstraintSystem, Error, Fixed, Instance, ProvingKey, VerifyingKey,
-    },
-    poly::{
-        commitment::{Params, ParamsProver},
-        kzg::{
-            commitment::{KZGCommitmentScheme, ParamsKZG},
-            multiopen::{ProverGWC, VerifierGWC},
-            strategy::AccumulatorStrategy,
-        },
-        Rotation, VerificationStrategy,
-    },
-    transcript::{TranscriptReadBuffer, TranscriptWriterBuffer},
+    load_grayscale_image, load_wnn,
 };
-use itertools::Itertools;
-use rand::{rngs::OsRng, RngCore};
-use snark_verifier::{
-    loader::evm::{self, encode_calldata, Address, EvmLoader, ExecutorBuilder},
-    pcs::kzg::{Gwc19, KzgAs},
-    system::halo2::{compile, transcript::evm::EvmTranscript, Config},
-    verifier::{self, SnarkVerifier},
-};
-use std::{env, fs, path::Path, rc::Rc};
-use zero_g::{checked_in_test_data::*, load_grayscale_image, load_wnn};
 
-type PlonkVerifier = verifier::plonk::PlonkVerifier<KzgAs<Bn256, Gwc19>>;
-
-#[derive(Clone, Copy)]
-struct StandardPlonkConfig {
-    a: Column<Advice>,
-    b: Column<Advice>,
-    c: Column<Advice>,
-    q_a: Column<Fixed>,
-    q_b: Column<Fixed>,
-    q_c: Column<Fixed>,
-    q_ab: Column<Fixed>,
-    constant: Column<Fixed>,
-    #[allow(dead_code)]
-    instance: Column<Instance>,
-}
-
-impl StandardPlonkConfig {
-    fn configure(meta: &mut ConstraintSystem<Fr>) -> Self {
-        let [a, b, c] = [(); 3].map(|_| meta.advice_column());
-        let [q_a, q_b, q_c, q_ab, constant] = [(); 5].map(|_| meta.fixed_column());
-        let instance = meta.instance_column();
-
-        [a, b, c].map(|column| meta.enable_equality(column));
-
-        meta.create_gate(
-            "q_a·a + q_b·b + q_c·c + q_ab·a·b + constant + instance = 0",
-            |meta| {
-                let [a, b, c] = [a, b, c].map(|column| meta.query_advice(column, Rotation::cur()));
-                let [q_a, q_b, q_c, q_ab, constant] = [q_a, q_b, q_c, q_ab, constant]
-                    .map(|column| meta.query_fixed(column, Rotation::cur()));
-                let instance = meta.query_instance(instance, Rotation::cur());
-                Some(
-                    q_a * a.clone()
-                        + q_b * b.clone()
-                        + q_c * c
-                        + q_ab * a * b
-                        + constant
-                        + instance,
-                )
-            },
-        );
-
-        StandardPlonkConfig {
-            a,
-            b,
-            c,
-            q_a,
-            q_b,
-            q_c,
-            q_ab,
-            constant,
-            instance,
-        }
-    }
-}
-
-#[derive(Clone, Default)]
-struct StandardPlonk(Fr);
-
-impl StandardPlonk {
-    fn rand<R: RngCore>(mut rng: R) -> Self {
-        Self(Fr::from(rng.next_u32() as u64))
-    }
-
-    fn instances(&self) -> Vec<Vec<Fr>> {
-        vec![vec![self.0]]
-    }
-}
-
-impl Circuit<Fr> for StandardPlonk {
-    type Config = StandardPlonkConfig;
-    type FloorPlanner = SimpleFloorPlanner;
-    type Params = ();
-
-    fn without_witnesses(&self) -> Self {
-        Self::default()
-    }
-
-    fn configure(meta: &mut ConstraintSystem<Fr>) -> Self::Config {
-        meta.set_minimum_degree(4);
-        StandardPlonkConfig::configure(meta)
-    }
-
-    fn synthesize(
-        &self,
-        config: Self::Config,
-        mut layouter: impl Layouter<Fr>,
-    ) -> Result<(), Error> {
-        layouter.assign_region(
-            || "",
-            |mut region| {
-                region.assign_advice(|| "", config.a, 0, || Value::known(self.0))?;
-                region.assign_fixed(|| "", config.q_a, 0, || Value::known(-Fr::ONE))?;
-
-                region.assign_advice(|| "", config.a, 1, || Value::known(-Fr::from(5)))?;
-                for (idx, column) in (1..).zip([
-                    config.q_a,
-                    config.q_b,
-                    config.q_c,
-                    config.q_ab,
-                    config.constant,
-                ]) {
-                    region.assign_fixed(|| "", column, 1, || Value::known(Fr::from(idx)))?;
-                }
-
-                let a = region.assign_advice(|| "", config.a, 2, || Value::known(Fr::ONE))?;
-                a.copy_advice(|| "", &mut region, config.b, 3)?;
-                a.copy_advice(|| "", &mut region, config.c, 4)?;
-
-                Ok(())
-            },
-        )
-    }
-}
-
-fn gen_srs(k: u32) -> ParamsKZG<Bn256> {
-    ParamsKZG::<Bn256>::setup(k, OsRng)
-}
-
-fn gen_pk<C: Circuit<Fr>>(params: &ParamsKZG<Bn256>, circuit: &C) -> ProvingKey<G1Affine> {
-    let vk = keygen_vk(params, circuit).unwrap();
-    keygen_pk(params, vk, circuit).unwrap()
-}
-
-fn gen_proof<C: Circuit<Fr>>(
-    params: &ParamsKZG<Bn256>,
-    pk: &ProvingKey<G1Affine>,
+async fn validate_evm<C: Circuit<Fr> + Clone>(
     circuit: C,
     instances: Vec<Vec<Fr>>,
-) -> Vec<u8> {
-    MockProver::run(params.k(), &circuit, instances.clone())
-        .unwrap()
-        .assert_satisfied();
-
-    let instances = instances
-        .iter()
-        .map(|instances| instances.as_slice())
-        .collect_vec();
-    let proof = {
-        let mut transcript = TranscriptWriterBuffer::<_, G1Affine, _>::init(Vec::new());
-        create_proof::<KZGCommitmentScheme<Bn256>, ProverGWC<_>, _, _, EvmTranscript<_, _, _, _>, _>(
-            params,
-            pk,
-            &[circuit],
-            &[instances.as_slice()],
-            OsRng,
-            &mut transcript,
-        )
-        .unwrap();
-        transcript.finalize()
-    };
-
-    let accept = {
-        let mut transcript = TranscriptReadBuffer::<_, G1Affine, _>::init(proof.as_slice());
-        VerificationStrategy::<_, VerifierGWC<_>>::finalize(
-            verify_proof::<_, VerifierGWC<_>, _, EvmTranscript<_, _, _, _>, _>(
-                params.verifier_params(),
-                pk.get_vk(),
-                AccumulatorStrategy::new(params.verifier_params()),
-                &[instances.as_slice()],
-                &mut transcript,
-            )
-            .unwrap(),
-        )
-    };
-    assert!(accept);
-
-    proof
-}
-
-fn gen_evm_verifier(
-    params: &ParamsKZG<Bn256>,
-    vk: &VerifyingKey<G1Affine>,
-    num_instance: Vec<usize>,
+    k: u32,
     name: &str,
-) -> Vec<u8> {
-    let protocol = compile(
-        params,
-        vk,
-        Config::kzg().with_num_instance(num_instance.clone()),
-    );
-
-    println!(
-        "Verification key: Number of fixed commitments: {}",
-        vk.fixed_commitments().len()
-    );
-
-    let vk = (params.get_g()[0], params.g2(), params.s_g2()).into();
-
-    let loader = EvmLoader::new::<Fq, Fr>();
-    let protocol = protocol.loaded(&loader);
-    let mut transcript = EvmTranscript::<_, Rc<EvmLoader>, _, _>::new(&loader);
-
-    let instances = transcript.load_instances(num_instance);
-    let proof = PlonkVerifier::read_proof(&vk, &protocol, &instances, &mut transcript).unwrap();
-    PlonkVerifier::verify(&vk, &protocol, &instances, &proof).unwrap();
-
-    let yul_code = loader.yul_code();
-
-    fs::write(format!("verifier_{name}.yul"), &yul_code).unwrap();
-
-    let bytecode = evm::compile_yul(&yul_code);
-
-    println!("Byte code size: {}", bytecode.len());
-
-    bytecode
-}
-
-fn evm_verify(deployment_code: Vec<u8>, instances: Vec<Vec<Fr>>, proof: Vec<u8>) {
-    let calldata = encode_calldata(&instances, &proof);
-    let success = {
-        let mut evm = ExecutorBuilder::default()
-            .with_gas_limit(u64::MAX.into())
-            .build();
-
-        let caller = Address::from_low_u64_be(0xfe);
-        let deployment_result = evm.deploy(caller, deployment_code.into(), 0.into());
-
-        let verifier = deployment_result.address;
-        match verifier {
-            Some(verifier) => {
-                let result = evm.call_raw(caller, verifier, calldata.into(), 0.into());
-
-                dbg!(result.gas_used);
-
-                !result.reverted
-            }
-            None => {
-                println!("Deployment result: {:?}", deployment_result.exit_reason);
-                false
-            }
-        }
-    };
-    assert!(success);
-}
-
-fn validate_evm<C: Circuit<Fr> + Clone>(circuit: C, instances: Vec<Vec<Fr>>, k: u32, name: &str) {
+    deploy: bool,
+) {
     println!("Generating Params...");
     let params = gen_srs(k);
     println!("Generating PK...");
@@ -277,11 +28,20 @@ fn validate_evm<C: Circuit<Fr> + Clone>(circuit: C, instances: Vec<Vec<Fr>>, k: 
 
     println!("Generating proof...");
     let proof = gen_proof(&params, &pk, circuit.clone(), instances.clone());
+
     println!("Verifying proof...");
-    evm_verify(deployment_code, instances.clone(), proof);
+    evm_verify(deployment_code.clone(), instances.clone(), proof.clone());
+
+    if deploy {
+        println!("Deploying...");
+        evm_deploy(deployment_code, instances.clone(), proof)
+            .await
+            .unwrap();
+    }
 }
 
-fn main() {
+#[tokio::main]
+async fn main() -> Result<()> {
     let args: Vec<String> = env::args().collect();
     if args[1] == "wnn" {
         let (k, model_path) = MNIST_MEDIUM;
@@ -292,14 +52,15 @@ fn main() {
 
         let outputs: Vec<Fr> = wnn.predict(&image).into_iter().map(Fr::from).collect();
         let instances = vec![outputs];
-        validate_evm(circuit, instances, k, &args[1])
+        validate_evm(circuit, instances, k, &args[1], args.len() > 2).await;
     } else if args[1] == "plonk" {
         let circuit = StandardPlonk::rand(OsRng);
         let instances = circuit.instances();
         let k = 8;
 
-        validate_evm(circuit, instances, k, &args[1])
+        validate_evm(circuit, instances, k, &args[1], args.len() > 2).await;
     } else {
         panic!("Unknown circuit: {:?}", args[1]);
     }
+    Ok(())
 }
